add app2 as mfe in container app
