+++ app2/src/components/AppTwo.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+// import logo from './logo.svg';
+// import './App.css';
+
+function AppTwo() {
+  return (
+    <div className="App">
+      <p>APP 2</p>
+      {/* <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.tsx</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header> */}
+    </div>
+  );
+}
+
+export default AppTwo;
+
