+++ app2/src/index.ts
@@ -1,0 +1,2 @@
+import('./bootstrap');
+//export {}
