+++ container/src/declarations.d.ts
@@ -2,3 +2,7 @@ declare module "app1/CounterAppOne" {
     const CounterAppOne: any;
     export default CounterAppOne;
 }
+declare module "app2/AppTwo" {
+    const AppTwo: any;
+    export default AppTwo;
+}
